+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -2,11 +2,11 @@ object Dependencies {
     const val material = "com.google.android.material:material:1.4.0-alpha01"
 
     object ComposeMaterialDialogs {
-        const val version = "0.3.2"
+        const val version = "0.3.3"
 
-        const val core = "com.vanpra.compose-material-dialogs:core:$version"
-        const val datetime = "com.vanpra.compose-material-dialogs:datetime:$version"
-        const val color = "com.vanpra.compose-material-dialogs:color:$version"
+        const val core = "io.github.vanpra.compose-material-dialogs:core:$version"
+        const val datetime = "io.github.vanpra.compose-material-dialogs:datetime:$version"
+        const val color = "io.github.vanpra.compose-material-dialogs:color:$version"
     }
 
     object Ktlint {
@@ -14,7 +14,7 @@ object Dependencies {
     }
 
     object Accompanist {
-        private const val version = "0.6.3.compose-7217131-SNAPSHOT"
+        private const val version = "0.7.1"
         const val pager = "com.google.accompanist:accompanist-pager:$version"
     }
 
@@ -27,10 +27,10 @@ object Dependencies {
     object AndroidX {
         const val appcompat = "androidx.appcompat:appcompat:1.2.0-rc01"
         const val coreKtx = "androidx.core:core-ktx:1.5.0-alpha03"
-        const val composeActivity = "androidx.activity:activity-compose:1.3.0-alpha04"
+        const val composeActivity = "androidx.activity:activity-compose:1.3.0-alpha06"
 
         object Compose {
-            const val version = "1.0.0-SNAPSHOT"
+            const val version = "1.0.0-beta04"
 
             const val ui = "androidx.compose.ui:ui:$version"
             const val material = "androidx.compose.material:material:$version"