+++ datetime/src/main/java/com/vanpra/composematerialdialogs/datetime/timepicker/TimePickerColors.kt
@@ -1,0 +1,79 @@
+package com.vanpra.composematerialdialogs.datetime.timepicker
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.material.SliderDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+/**
+ * Represents the colors used by a [timepicker] and its parts in different states
+ *
+ * See [TimePickerDefaults.colors] for the default implementation
+ */
+interface TimePickerColors {
+    val border: BorderStroke
+
+    /**
+     * Gets the background color dependant on if the item is active or not
+     *
+     * @param active true if the component/item is selected and false otherwise
+     * @return background color as a State
+     */
+    @Composable
+    fun backgroundColor(active: Boolean): State<Color>
+
+    /**
+     * Gets the text color dependant on if the item is active or not
+     *
+     * @param active true if the component/item is selected and false otherwise
+     * @return text color as a State
+     */
+    @Composable
+    fun textColor(active: Boolean): State<Color>
+
+    fun selectorColor(): Color
+    fun selectorTextColor(): Color
+
+    @Composable
+    fun periodBackgroundColor(active: Boolean): State<Color>
+}
+
+internal class DefaultTimePickerColors(
+    private val activeBackgroundColor: Color,
+    private val inactiveBackgroundColor: Color,
+    private val activeTextColor: Color,
+    private val inactiveTextColor: Color,
+    private val inactivePeriodBackground: Color,
+    private val selectorColor: Color,
+    private val selectorTextColor: Color,
+    borderColor: Color
+) : TimePickerColors {
+    override val border = BorderStroke(1.dp, borderColor)
+
+    @Composable
+    override fun backgroundColor(active: Boolean): State<Color> {
+        SliderDefaults
+        return rememberUpdatedState(if (active) activeBackgroundColor else inactiveBackgroundColor)
+    }
+
+    @Composable
+    override fun textColor(active: Boolean): State<Color> {
+        return rememberUpdatedState(if (active) activeTextColor else inactiveTextColor)
+    }
+
+    override fun selectorColor(): Color {
+        return selectorColor
+    }
+
+    override fun selectorTextColor(): Color {
+        return selectorTextColor
+    }
+
+    @Composable
+    override fun periodBackgroundColor(active: Boolean): State<Color> {
+        return rememberUpdatedState(if (active) activeBackgroundColor else inactivePeriodBackground)
+    }
+}